Extract context create function to a factory file
